Resolved requested changes v2.
